Filter transform when view box changes (#607)
* Chore: Made some changes to fix the issue. See description.
in table.component.ts, the assigning for filters is now done by creating a new map with the filters in it. This is done so the active-filter component is properly updated.
in table.component.ts the metadataLabels were passed to the filterService but that should be metadataNames.
filters are now transformed to new state if view box is changed where some columns do not exists. It deletes the values from the map.
Added new cypress test for previous statement.
Changed [] to Array<{type}> for better readability and conventional coding.

* Chore: optimising the code a little bit.
Instead of filterContextSubject being updated for every changed column. It is now only changed once if any column was changed.

* Chore: removed unnecessary type assigning for map.

* chore: Removed new map assigning from changeFilter and ordered the class variables correctly in filter service